#10303 TESTS_RAN:Patrick's manual test
Description of the problem:
This problem is caused by one or more errors in the striping logic
in mcd_fth_aio_blk_read() and mcd_fth_aio_blk_write().
For some reason, this logic tries to stripe reads/writes across
multiple stripes even if there is only one stripe!

The stripe size (Mcd_aio_strip_size) is
assigned to the segment size divided by the number of files
when the slab subsystem is initialized (mcd_osd_slab_init()).

I think this error can happen for both FIFO and slab containers.

With FIFO mode containers, many writes are collected in a segment
sized buffer and written out in a single large (segment sized?)
write.  I suspect that this write is always nicely aligned and
sized so that it would not induce striping problems.
FIFO reads, however, are made per object, and an object may
be arbitrarily positioned within the large write buffer.
Objects may therefore cross the strip boundary.
The likelihood of crossing a strip boundary is highest for
large objects.

With slab
mode containers place objects in slabs that are nicely aligned
within a segment.  However, the stripe size is set to the size
of a segment divided by the number of flash files, so it could
be less than the 8MB max object size.  This could cause strip
error problems.

I believe the 2 scenarios we saw at Morningstar are for FIFO
containers because they happened for reads.  A strip error for
a slab mode container could only occur for writes (I think!).
Nick checked and Morningstar uses lots of FIFO mode containers.



git-svn-id: https://10.196.60.217/svn/schooner-products/branches/membrain-4.3/membrain/server@1493 01a69087-22d5-4a29-8152-8a1d5e10e5e9