#11005 Btree becomes inconsitent after write fails due to out of space TESTS_RAN: (1) Dev test(629) (2) Cassandra Performance test(5M rows 10c 10B) (3) The following tests were run and verified    (a) check if writes fail with out of storage error after softlimit reached    (c) check of flash space allocated counter at Btree layer  is correct after restart    (d) check if the flash space allocated counter at btree layer matches with FDFStats    (e) check if proper message is printed when out of space occurs    (f) check if soft limit can be disabled or enabled through FDF_DISABLE_SOFT_LIMIT_CHECK    (g) check if soft limit can be configured through property FDF_FLASH_SIZE_SOFT_LIMIT    (h) check if writes can succeed after space reclaimed
Issue:
When apps writes object, it might result in to modification of one or more btree nodes.
The modified nodes are written to flash independnetly after modiefied in memory. The writes  of one or more btree nodes might fails if storage becomes full.
The expected operation is to rollback all the modified nodes and fail the application write operation. But 2.0 release does not support runtime rollback .
So, the partially written nodes are not rolledback and we written failure to application.  The btree becomes inconsistent and subsequent operations causes aborts.

Fix:
Ideal fix is to implement rollback. This will be done in the future releases.
The following fix/workaround will be used in 2.0.
This checkin introduces softlimit over the total flash space. All the write operations check if this limit is reached before proceeding with a write.
The softlimit is configured is configured such a way that, there is enough room for pending writes and write failures due to space issue does not occur.
By default the softlimit is set to 85% of Flash space.  But this can be changed through FDF_FLASH_SIZE_SOFT_LIMIT variable. The soft limit check is enabled by default.
When softlimit reached, the write call fails with FDF_OUT_OF_STORAGE_SPACE error.  App needs to delete some objects for reclaiming space.

Configuration:
FDF_FLASH_SIZE_SOFT_LIMIT = <size in GB> - Flash size soft limit. If not specified, 85% of flash space is set as softlimit
FDF_DISABLE_SOFT_LIMIT_CHECK = < 0 or 1 > - Disable the softlimit check



git-svn-id: https://10.196.60.217/svn/schooner-products/branches/fdf-2.0@2121 01a69087-22d5-4a29-8152-8a1d5e10e5e9