#11427 FDF aborts with assert `dry_run || new_used_space <= bt->nodesize_less_hdr' in a stress test TESTS_RAN: Devtest, Unit tests and long run test(~10hrs) Code Review done by Hari & EF. Checking in on-behalf of Ramesh.
Issue:
The Merge operation,during btree rebalance, copies objects to left and write nodes.
During this copy operation, space optimization logic asserts reporting that not enough room in the node.
This is an assert introduced in the initial checkin to catch these kind of conditions. 

When we copy the objects to the left node, it goes through prefix optimization logic insterad of just plain copy. 
In some rare cases, the actual space required to hold the chunk of objects increases little bit because common prefix might change after copy. so the space required goes beyond actual space available. The assert catches these kind of situations and aborts to avoid memory corruptions.

Fix:
The fix is to check first if there would enough space to hold if prefix optimization applied without actually applying directly and proceed with merge if there is enough space.  If not, skip the merge and balance the tree with equalize key function. 


git-svn-id: https://10.196.60.217/svn/schooner-products/branches/fdf-2.0@2145 01a69087-22d5-4a29-8152-8a1d5e10e5e9