#11463 FDF Log messages do not show up in log file TESTS_RAN: Dev Test and unit tests
In the current code, logs writtent to stderr in btree layer is redirected to fdf log. But cassandra redirects stdrr to /dev/null, so all the btree layer messages are lost. 
The fix adds a call back for the btree layer to log message in to the FDF log file. 

Additional Tests RUN:
(1) Check if btree layer messages are printed in fdf log file during cassandra test
(2) Check if btree layer messages are printed in stderr if log file is not specified 
(3) Check if logging and funcionality works as expected without btree layer
(4) check if changes causes any memory leak


git-svn-id: https://10.196.60.217/svn/schooner-products/branches/fdf-2.0.10@2181 01a69087-22d5-4a29-8152-8a1d5e10e5e9