#11529 FDFDelete container after recovery needs more than 1 hr in daily QA The problem is with incorrect durability setting applied to VDC shard after session restart. By default, each physical shard has FULL durability. A logical container passes its durability to shard and it is used by lower layers such as logging. For a FULL durable shard, log record of a deleted object is written with fsync call which is blocking and causes delays in deletion call path.
The current wrong behavior of durability of a container becoming visible to other
containers needs a fix. It should be strictly private to a container and shard should not
change its durability at will.

The current fix for the bug do not temper exiting code and before deleting objects of a 
container, set shard durability to container's durability.
TESTS_RAN: Dev test, unit test


git-svn-id: https://10.196.60.217/svn/schooner-products/branches/fdf-2.0.10@2240 01a69087-22d5-4a29-8152-8a1d5e10e5e9