#FDF-153 TESTS_RAN: Dev test t/Zsck/Z15_metafault_flog_recovery_64K.t
Issue:
Flog corruption introdced by error injector(zsmetafault) not detected by ZS during restart.
Reason and Fix:
with the fix for the flog corruption issue(#FDF-194), up to two log blocks can be safely discorded and abort in the event of log corruption is optional. 
Zsmeta fault just corrupted 1K in the flog so we safely discarded and continued successful restart. So this is not actually a bug. 
Just to make sure that the zsck test cases work, I have changed the zsmetafault to corrupt 3 log blocks. With ZS_LOG_ABORT_ON_CORRUPTION=1, the test case for 153 should succeed.






git-svn-id: https://10.196.60.217/svn/schooner-products/branches/fdf-3.0@2582 01a69087-22d5-4a29-8152-8a1d5e10e5e9